Rename theme-options.php to class-okfnthemeoptions.php and update import
